Пример исправления
