+++ OOp/Stroka.h
@@ -1,2 +1,5 @@
 #pragma once
+
+#include "Iter.h"
+
 void strOut();
